pyproject.toml
[tool.poetry]
name = "pii_recognition"
version = "0.1.0"
description = ""
authors = ["Gabe <wei.g.chu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
spacy = "^2.2.3"
nltk = "^3.4.5"
scikit-learn = "^0.22.2"
numpy = "^1.18.1"
python-crfsuite = "^0.9.6"
mlflow = "^1.7.0"
stanza = "^1.0.0"
dagster = "^0.7.9"
pandas = "^1.0.3"
pakkr = "^0.1.0"
python-decouple = "^3.3"
flair = "^0.6.1"
boto3 = "^1.16.0"
google-cloud-language = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.7.9"
mypy = "^0.761"
isort = "^4.3.21"
pytest-env = "^0.6.2"
moto = "^1.3.16"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[mypy]
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
